Moved pyosys generator script
The pyosys generator script was moved to it's own directory so that
there is a suitable path to store pyosys specific documentation.

'pyosys_src' was chosen as dirname over the obvious choice of 'pyosys'
because the second one would interfere with the pyosys package name, so
if one would try to build yosys with python while pyosys is already
installed in the system, using pyosys as dirname would result in the
build-process to look for the generator script in the install-directory.